requirements.txt
# Core dependencies com versões fixas
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
aiohttp==3.9.1
supabase==2.3.0

# Security
cryptography==41.0.7
safety==3.0.1

# Performance e Rate Limiting
aiometer==0.5.0
